% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\title{Bayesian information criterion}
\usage{
bic(x, output)
}
\arguments{
\item{x}{the data set}

\item{output}{The output from the \code{mcmc} function}
}
\description{
Function for calculating Bayesian information criterion (BIC)
for a single model. The formula is -2*log-like + npar*log(n),
where n is the sample size.
}

