% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernels.R
\name{exp_cdf}
\alias{exp_cdf}
\title{Unit exponential CDF}
\usage{
exp_cdf(theta)
}
\arguments{
\item{theta}{Should be positive}
}
\description{
The Unit exponential CDF function. Used as a kernel in the \code{mcmc} function.
}

