% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_xmins.R
\name{get_xmins}
\alias{get_xmins}
\title{Xmin inference}
\usage{
get_xmins(thetas, x, upper = 0.99)
}
\arguments{
\item{thetas}{Matrix of parameters. Typically posterior samples}

\item{x}{Values of x to scan}

\item{upper}{arg min cut-off. Default 0.99}
}
\description{
Xmin inference
}

