% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initialise_pars.R
\name{initialise_pars}
\alias{initialise_pars}
\title{Initialise parameter values for the MCMC algorithm}
\usage{
initialise_pars(pars, type = c("displ", "dislnorm", "disexp", "diff",
  "difflnorm", "diffsq", "diffxmax", "difflnormxmax", "dislnormpl",
  "pairdiffpl"))
}
\arguments{
\item{pars}{A vector of initial parameter values}

\item{type}{A string indicating the model to be fitted}
}
\description{
Initialise parameter values for the MCMC algorithm
}

