% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lines.R
\name{lines.plmcmc_mat}
\alias{lines.diff}
\alias{lines.difflnorm}
\alias{lines.dislnorm}
\alias{lines.dislnormpl}
\alias{lines.displ}
\alias{lines.plmcmc_mat}
\title{Fitted lines}
\usage{
\method{lines}{plmcmc_mat}(x, xmax = 10000, n = NULL, ...)

\method{lines}{diff}(x, xmax = 10000, n, ...)

\method{lines}{difflnorm}(x, xmax, n, ...)

\method{lines}{dislnorm}(x, xmax, n, ...)

\method{lines}{displ}(x, xmax, n, ...)

\method{lines}{dislnormpl}(x, xmax, n, ...)
}
\arguments{
\item{x}{Output from the \code{mcmc} function}

\item{xmax}{Maximum x value to plot}

\item{n}{The number of samples from \code{x} to plot}

\item{...}{Additional lines arguments}
}
\description{
Adds the fitted to a plot
}

