% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcmc}
\alias{mcmc}
\title{Main MCMC function}
\usage{
mcmc(pars, cov_mat, data, N, thin = 1, kern = NULL, verbose = TRUE, ...)
}
\arguments{
\item{pars}{Initial parameter values}

\item{cov_mat}{variance-covariance matrix used in the Normal random walk}

\item{data}{The data set (a vector)}

\item{N}{Number of MCMC iterations}

\item{thin}{Should the MCMC output be thinned}

\item{kern}{default \code{NULL}. The difference kernel to be used. When fitting a standard log normal or
power law distribution, we don't need a difference kernel. Otherwise, a function should be passed. For example,
\code{exp_cdf}}

\item{verbose}{default \code{TRUE}. On verbose mode, 1 in 100 iterations and the acceptance
rate will be printed to the screen.}
}
\value{
A matrix
}
\description{
Function used to perform parameter inference
}

