% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Log-log plot of the data}
\usage{
plot_fit(x, ...)
}
\arguments{
\item{x}{A vector of data values}

\item{...}{Additional arguments passed to the plot function}
}
\description{
Log-log plot of the data
}

