% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_mcmc.R
\name{plot_mcmc}
\alias{plot_mcmc}
\title{Plot of MCMC output}
\usage{
plot_mcmc(output, ...)
}
\arguments{
\item{output}{A matrix obtained from the \code{mcmc} method.}

\item{...}{Additional arguments}
}
\description{
Plot of MCMC output
}

