% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proposals.R
\name{update_parameters}
\alias{update_parameters}
\title{Parameter proposals}
\usage{
update_parameters(pars, cov_mat)
}
\arguments{
\item{pars}{model parameters}

\item{cov_mat}{covariance matrix}
}
\description{
Parameter proposals
}

